feat: create hook for categories fetching
